change permissions from tags to capabilities in savedQueries

+++ x-pack/plugins/osquery/server/routes/saved_query/delete_saved_query_route.ts
@@ -6,11 +6,11 @@
  */
 
 import { schema } from '@kbn/config-schema';
-import { PLUGIN_ID } from '../../../common';
 import { IRouter } from '../../../../../../src/core/server';
 import { savedQuerySavedObjectType } from '../../../common/types';
+import { OsqueryAppContext } from '../../lib/osquery_app_context_services';
 
-export const deleteSavedQueryRoute = (router: IRouter) => {
+export const deleteSavedQueryRoute = (router: IRouter, osqueryContext: OsqueryAppContext) => {
   router.delete(
     {
       path: '/internal/osquery/saved_query/{id}',
@@ -19,9 +19,17 @@ export const deleteSavedQueryRoute = (router: IRouter) => {
           id: schema.string(),
         }),
       },
-      options: { tags: [`access:${PLUGIN_ID}-writeSavedQueries`] },
     },
     async (context, request, response) => {
+      const [coreStartServices] = await osqueryContext.getStartServices();
+
+      const {
+        osquery: { writeSavedQueries },
+      } = await coreStartServices.capabilities.resolveCapabilities(request);
+
+      if (!writeSavedQueries) {
+        return response.forbidden();
+      }
       const savedObjectsClient = context.core.savedObjects.client;
 
       await savedObjectsClient.delete(savedQuerySavedObjectType, request.params.id, {
